Fix: Resolve production HTTP 500.30 by correcting SessionId type mismatches
- Root Cause: Entity Framework model validation failed due to foreign key type incompatibility
  - AuditLog.SessionId and Issue.SessionId were long? (INT64)
  - Session.SessionId primary key is int (INT32)
  - EF Core threw InvalidOperationException during DbContext initialization

- Changed SessionId from long/long? to int/int? across 9 files:
  - Models: AuditLog.cs (line 17), Issue.cs (line 30)
  - Controllers: IssueController.cs (CreateIssueRequest line 203), AdminController.cs (line 136), HostController.cs (lines 363, 403, 1340)
  - Hubs: SessionHub.cs (9 methods), QAHub.cs (2 methods)
  - Services: AssetProcessingService.cs (line 38), UnifiedHtmlTransformService.cs (line 34)

- Enhanced Program.cs database validation:
  - Changed fallback connections from Trusted_Connection=true to SQL Auth (sa account)
  - Added comprehensive NOOR-DEBUG logging to capture startup errors
  - Added inner exception logging for detailed error diagnosis
  - Lines 65-79: Updated all DbContext fallback connection strings
  - Lines 330-408: Added enhanced validation logging

- Simplified appsettings configuration:
  - Removed duplicate connection strings (SimplifiedConnection, KSessionsDb, KQurDb)
  - Standardized to single DefaultConnection -> KSESSIONS for production
  - All DbContexts now use same connection with SQL Authentication

Production Validation Results:
- Website: HTTP 200 OK at http://noorcanvas.servehttp.com
- Database: Both CanvasDbContext and KSessionsDbContext connected successfully
- Logs: 0 critical errors, 0 warnings
- Build: 0 compilation errors (3 non-critical style warnings)

Deployment: 2025-10-12_13-28-03 via ncdeploy.ps1

+++ SPA/NoorCanvas/Controllers/AdminController.cs
@@ -133,7 +133,7 @@ namespace NoorCanvas.Controllers
         }
 
         [HttpPost("session/{sessionId}/terminate")]
-        public async Task<IActionResult> TerminateSession(long sessionId, [FromQuery] string sessionToken)
+        public async Task<IActionResult> TerminateSession(int sessionId, [FromQuery] string sessionToken)
         {
             try
             {
